# app/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/link/{publicLinkId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "publicLinkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Download By Public Link",
        "tags" : [ "Public Blob Resource" ]
      }
    },
...
    "/public/download-multiple" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "type" : "string",
                "format" : "binary"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        },
        "summary" : "Download Public Multiple",
        "tags" : [ "Public Blob Resource" ]
      }
    },
    "/public/download-multiple/check" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "format" : "binary"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Check Download Public Multiple",
        "tags" : [ "Public Blob Resource" ]
      }
    },
    "/public/download/{nodeId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "access_code",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "node_link_id",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Download Public File",
        "tags" : [ "Public Blob Resource" ]
      }
    },
    "/public/download/{nodeId}/check" : {
      "get" : {
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "access_code",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "node_link_id",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Check Download Public File",
        "tags" : [ "Public Blob Resource" ]
      }
    },
    "/public/link/download/{publicLinkId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "publicLinkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Download Via Public Link",
        "tags" : [ "Public Blob Resource" ]
      }
    },
...
  },
...
  "info" : {
    "title" : "carbonio-files-ce-app API",
    "version" : "unversioned"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}